requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
twilio
python-multipart
google-generativeai
